requirements.txt
# Core dependencies
requests>=2.31.0

# Optional: SignalR client for full pub/sub testing
signalrcore>=0.9.5

# Azure Functions dependencies (for local development)
azure-functions>=1.17.0
azure-functions-worker>=1.19.0